+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-17T12:53:05.298Z",
+    "2026-05-18T05:04:23.600Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {